refactor: clarify module names and imports
